feat(api): Implement mask for joints
- A masked joint behaves like a normal joint, but all read/write operations
  are silently discarded while still returning success.
- This API can be used when certain joints are disconnected or damaged.

+++ include/wujihandcpp/data/helper.hpp
@@ -24,7 +24,7 @@ struct ReadOnlyData {
 
     using ValueType = ValueType_;
 
-    static constexpr StorageInfo info(uint64_t) {
+    static constexpr StorageInfo info(uint32_t) {
         return StorageInfo{sizeof(ValueType), index, sub_index, 0};
     }
 }; // namespace data
@@ -43,7 +43,7 @@ struct WriteOnlyData {
 
     using ValueType = ValueType_;
 
-    static constexpr StorageInfo info(uint64_t) {
+    static constexpr StorageInfo info(uint32_t) {
         return StorageInfo{sizeof(ValueType), index, sub_index, 0};
     }
 };
